clickbankforge/scraper.py
import json
import os
from typing import List, Dict
import requests
from bs4 import BeautifulSoup

def load_top_clickbank_products(fallback_path: str) -> List[Dict]:
    """
    Load top ClickBank products, either from live scraping or fallback file.
    
    Args:
        fallback_path: Path to JSON file with fallback product data
        
    Returns:
        List of product dictionaries with offer details
    """
    try:
        # Try to scrape live data first
        products = scrape_clickbank_marketplace()
        if products:
            return products
    except Exception as e:
        print(f"Scraping failed: {e}. Using fallback data.")
    
    # Fallback to static data
    with open(fallback_path, 'r') as f:
        return json.load(f)

def scrape_clickbank_marketplace() -> List[Dict]:
    """
    Scrape the ClickBank marketplace for top products.
    
    Returns:
        List of product dictionaries with offer details
    """
    # TODO: Implement actual ClickBank API integration
    # This is a placeholder that returns empty list
    return [] 