requirements.txt
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.0
pandas==2.1.0
python-telegram-bot==20.6
notion-client==2.2.0
beautifulsoup4==4.12.2
lxml==4.9.3 